requirements.txt
flask
psycopg2
hashlib
pyjwt
datetime
random
os